+++ lionengine-core/src/test/java/com/b3dgs/lionengine/core/sequence/FrameTest.java
@@ -49,7 +49,7 @@ public class FrameTest
             }
 
             @Override
-            public void computeFrameRate(double lastTime, double currentTime)
+            public void computeFrameRate(long lastTime, long currentTime)
             {
                 result.set(true);
             }
